Remove filename entropy algorithm.
This will never actually work. In the case of a single producer which
handles more than one file in a millisecond, the second temp file is
created after the first file is moved, so there won't be a conflict -
just a silent overwrite on the second file move. There's a similar race
condition when dealing with two producers creating files simultaneously.

Instead, just work using uuids, either provided or generated.